fix(Linear Node): Correct Release stage selection and Customer Need priority type
Live verification against the Linear API surfaced two schema mismatches:
- Release.stage is an object (ReleaseStage), so select id/name subfields
- CustomerNeed priority is a Float, not an Int

+++ packages/nodes-base/nodes/Linear/v2/actions/customerNeed/create.operation.ts
@@ -73,7 +73,7 @@ export async function execute(
 			const additionalFields = this.getNodeParameter('additionalFields', i) as IDataObject;
 
 			const body = {
-				query: `mutation CustomerNeedCreate($customerId: String!, $body: String, $issueId: String, $projectId: String, $priority: Int) {
+				query: `mutation CustomerNeedCreate($customerId: String!, $body: String, $issueId: String, $projectId: String, $priority: Float) {
 					customerNeedCreate(input: {
 						customerId: $customerId
 						body: $body
